requirements.txt
disnake
motor
beanie
aiofiles
aiohttp
python-dotenv
loguru
ruff